requirements.txt
ipykernel
pandas
matplotlib
numpy
tensorflow
tensorflow-hub
tf_keras
streamlit
pillow